Rename cmd_mods to cmds
